More clippy warnings fixed manually.

+++ support/teeny-vec/benches/teeny_vec.rs
@@ -12,12 +12,12 @@ fn bench_push_small(c: &mut Criterion) {
     let mut group = c.benchmark_group("push_small_inline");
 
     // Small sizes that fit inline (14 bytes for TeenyVec)
-    for size in [1, 4, 8, 12, 14] {
+    for size in [1_u8, 4, 8, 12, 14] {
         group.bench_with_input(BenchmarkId::new("TeenyVec", size), &size, |b, &size| {
             b.iter(|| {
                 let mut vec = TeenyVec::new();
                 for i in 0..size {
-                    vec.push(black_box(i as u8));
+                    vec.push(black_box(i));
                 }
                 black_box(vec);
             });
@@ -27,7 +27,7 @@ fn bench_push_small(c: &mut Criterion) {
             b.iter(|| {
                 let mut vec = SmallVec::<[u8; 16]>::new();
                 for i in 0..size {
-                    vec.push(black_box(i as u8));
+                    vec.push(black_box(i));
                 }
                 black_box(vec);
             });
@@ -37,7 +37,7 @@ fn bench_push_small(c: &mut Criterion) {
             b.iter(|| {
                 let mut vec = Vec::<u8>::new();
                 for i in 0..size {
-                    vec.push(black_box(i as u8));
+                    vec.push(black_box(i));
                 }
                 black_box(vec);
             });
@@ -51,12 +51,12 @@ fn bench_push_medium(c: &mut Criterion) {
     let mut group = c.benchmark_group("push_medium_heap");
 
     // Medium sizes that require heap allocation
-    for size in [20, 32, 64, 128] {
+    for size in [20_u8, 32, 64, 128] {
         group.bench_with_input(BenchmarkId::new("TeenyVec", size), &size, |b, &size| {
             b.iter(|| {
                 let mut vec = TeenyVec::new();
                 for i in 0..size {
-                    vec.push(black_box(i as u8));
+                    vec.push(black_box(i));
                 }
                 black_box(vec);
             });
@@ -66,7 +66,7 @@ fn bench_push_medium(c: &mut Criterion) {
             b.iter(|| {
                 let mut vec = SmallVec::<[u8; 16]>::new();
                 for i in 0..size {
-                    vec.push(black_box(i as u8));
+                    vec.push(black_box(i));
                 }
                 black_box(vec);
             });
@@ -76,7 +76,7 @@ fn bench_push_medium(c: &mut Criterion) {
             b.iter(|| {
                 let mut vec = Vec::<u8>::new();
                 for i in 0..size {
-                    vec.push(black_box(i as u8));
+                    vec.push(black_box(i));
                 }
                 black_box(vec);
             });
